refactor(dir): Adicionado pastas para controller e service
