+++ src/client/entities/client.entity.ts
